requirements.txt
browser_cookie3==0.19.1
PyYAML==6.0.1
Requests==2.31.0
